% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearNegImage.R
\docType{data}
\name{clearNegImage}
\alias{clearNegImage}
\title{A negative red-blue non RGB TIFF image with low amounts of noise}
\format{
A list of two matrices, the first red (AF568), 
the second blue (DAPI).
}
\usage{
data("clearNegImage")
}
\description{
This is a negative control image, which is one of three images
meant to be used for testing purposes in the package.
}
\keyword{datasets}
