% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getImageStats.R
\name{getImageStats}
\alias{getImageStats}
\title{Generation of fluorescence-related scores.}
\usage{
getImageStats(
  imgDirs,
  imgNames,
  frameNumFocus,
  frameNumReference = FALSE,
  sizeCutoff,
  diagnoImgs = FALSE,
  outDir = ".",
  numPix = "All",
  highNoise = FALSE,
  intensityCutoffFocus = TRUE,
  intensityCutoffReference = TRUE,
  reportIntensity = FALSE,
  nuclearToCellQuotient = 0.66,
  numOfImgs = "All"
)
}
\arguments{
\item{imgDirs}{A vector or list of pathways, including filenames, to the
images to be analysed, e.g. "Raw_images/Positive_ctrl.nd2". Formats that are 
currently supported are nd2, czi, tiff, png or lists of images, in the form 
of three-dimensional arrays, where each layer in the third dimension 
represents a color. nd2, czi or non-normalised, integer TIFF are clearly 
preferable for memory and resolution purposes.}

\item{imgNames}{The names of the images. This argument is used both for
naming of the rows in the output matrix, but also for naming of the images
if such are generated using the diagnoImgs flag.}

\item{frameNumFocus}{This identifies which of the frames in the files that 
contains the information that should be measured primarily, such as IgG
binding in the case of autoantibody screening.}

\item{frameNumReference}{Optionally, this identifies a frame in the files 
that can be used to focus the analysis to certain regions. It could be
that only areas with GFP expression can contain meaningful expression of 
the marker of primary interest, and then only these areas will be considered.}

\item{sizeCutoff}{The size of a typical cell nucleus. This is used to 
identify structures of a size equal to or larger than this cell threshold.
Smaller objects that are ring-shaped will be excluded with this cutoff, which
reduces noise considerably, but the threshold should not be set too high as
it then risks leading to the exclusion of true positive cells. This value
can be generated using the \code{\link{sizeCutoffGen}} function.}

\item{diagnoImgs}{Should a images delimiting the islands that have been
selected be returned?}

\item{outDir}{The directory that the diagnoImages should be saved in. Only
used together with diagnoImgs = TRUE.}

\item{numPix}{If the frames are very large, this can be used to reduce the 
computational burden. NB! If this command is used, then the images
will not be identical from round to round, and will not be comparable to the
output of other functions, as a random subset of the pictures are used.}

\item{highNoise}{Is the assay especially noisy, with very high and individual
background? If so, a second run of noise reduction can be used to increase
signal-to-noise ratio. This is often but not always needed for fixed 
cell-based assays, e.g.}

\item{intensityCutoffFocus}{If an external intensity cutoff is used, e.g. if a 
large number of images from the same series are used, one can save
computational time by defining the background on the positive control. See
\code{\link{intensityCutoffGen}}. The two possible values here are a number
and "TRUE", in which each individual frame gets a separate intensity cutoff.
This latter option is suitable when it is unclear if all samples in a series
have been generated with the same imaging settings, or, more commonly, if RGB-
compliant files are used as input, as these are always normalised to a 
range from 0 to 1, and thus, the background in a negative frame will have 
considerably higher values than in a positive sample, simply because the 
rate from the lowest to the highest signal is more compressed in a case 
with noise only. If individual values are used, thresholding is performed 
with the auto_threshold function from the autothresholdr package, that uses
the same thresholding algorithms as imageJ. Specifically, it is the "tri"
option that is used. 
See: \url{https://www.example.com}}

\item{intensityCutoffReference}{Same as above but for the control frame,
if present.}

\item{reportIntensity}{Should the sum of the intensities for all the 
surviving islands be returned? Default is FALSE.}

\item{nuclearToCellQuotient}{The quotient of the median nuclear width divided
by the median cell width. Set to 0.66, as this is the quotient for HEK cells
according to the literature. This is used to calculate how many cells that
are clumped together in large, confluent multicell objects in the pictures.}

\item{numOfImgs}{If the provided files are nd2 format, they can contain
multiple files. In this case, this flag can be used to restrict the number of
used images.}
}
\value{
A data frame with a number of statistics for the individual images.
Optionally, a directory containing images identifying the regions that have
been identified as cells of interest.
}
\description{
This function is the central umbrella function that takes a list of image 
files or image file directories and generates a list of statistics for each
of the files.
}
\examples{
#Retrieve the example data
data(posImage)
data(negImage)

sizeCutoff <- getSizeCutoff(imgDirs = list(negImage), frameNum = 3)
intensityCutoff <- getIntensityCutoff(imgDirs = list(negImage, posImage), 
frameNum = 1)

#And now, for the final generation of results
result <- getImageStats(imgDirs = allImgs,
                       imgNames =c("Pos", "Noisy_neg", "Clear_neg"),
                       frameNum = 1,
                       sizeCutoff = sizeCutoff,
                       intensityCutoff = intensityCutoff)
                       
#And here an example of how to use it with generation of images. 

result <- getImageStats(imgDirs = allImgs,
                       imgNames =c("Pos", "Noisy_neg", "Clear_neg"),
                       frameNum = 1,
                       sizeCutoff = sizeCutoff,
                       diagnoImgs = TRUE,
                       outDir = ".", 
                       intensityCutoff = intensityCutoff)
}
\seealso{
\code{\link[autothresholdr]{auto_thresh}}
}
