% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intensityCutoffGen.R
\name{intensityCutoffGen}
\alias{intensityCutoffGen}
\title{Identification of an intensity cutoff for a whole experiment}
\usage{
intensityCutoffGen(imgDirs, frameNum, numOfImgs = "All", numPix = "All")
}
\arguments{
\item{imgDirs}{The directory where the file(s) that will be used to generate
the threshold is kept. If more than one file, then a median value is
generated. Currently, this function is only implemented to work with 
raw-value non-RGB TIFF files. 
There is an option here, which is mainly for example purposes, using the
provided dataset, and that is that a list of image files can be provided 
directly. These should then be in the form of a list of lists, each sublist 
being composed of matrices representing the colors. See example for useage.}

\item{frameNum}{This identifies which of the frames in the file that 
contains the information about the autoantibody binding.}

\item{numOfImgs}{If the provided files are nd2 format, they can contain
multiple files. In this case, this flag can be used to restrict the number of
used images.}

\item{numPix}{If the frames are very large, this can be used to reduce the 
computational burden. For reproducibility reasons, it might be clever to
run this multiple times, in this case, to bootstrap, or alternatively to
set a seed before starting.}
}
\value{
An intensity cutoff value.
}
\description{
With this function, the otherwise integrated intensity cutoff is 
externalised. This is ONLY useful in situations where the original values
from the imaging have not been perturbed, so that the individual frames have
different max values, and are not all compressed to a range from 0 to 1. NB!
This requirement is not fulfilled if the files are exported as RGB compliant,
as the RGB format always ranges from 0 to 1.
}
\examples{
See example in \code{\link{imageStatGen}}
}
