% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noisyNegImage.R
\docType{data}
\name{noisyNegImage}
\alias{noisyNegImage}
\title{A negative red-blue non RGB TIFF image with high, but normal amounts of noise}
\format{
A list of two matrices, the first red (AF568), 
the second blue (DAPI).
}
\usage{
data("noisyNegImage")
}
\description{
This is a negative control image, which is one of three images
meant to be used for testing purposes in the package.
}
\keyword{datasets}
