% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posImage.R
\docType{data}
\name{posImage}
\alias{posImage}
\title{A few CASPR2-positive cells}
\format{
An a x b x c array representing red (IgG), green (GFP) and
blue (DAPI) colors.
}
\usage{
data("posImage")
}
\description{
This is a positive control image, used for help functions and vignettes. It
is a real-life image showing CASPR2 expressing HEK293-MSC cells co-incubated
with CASPR2-IgG. It was generated by Dr Amanda Freitas Huhtamäki during her
postdoctoral period in the Jakob Theorell team, Center for Infectious 
Medicine, Department of Medicine Huddinge, in the Live Cell Imaging core 
facility, Huddinge, both Karolinska Institutet, Sweden.
}
\keyword{datasets}
