% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sizeCutoffGen.R
\name{sizeCutoffGen}
\alias{sizeCutoffGen}
\title{Identification of a suitable size cutoff for cells}
\usage{
sizeCutoffGen(
  imgDirs,
  frameNum,
  returnAllVals = FALSE,
  numPix = "All",
  numOfImgs = "All"
)
}
\arguments{
\item{imgDirs}{The directory containing the image file/files that should be
used to identify the cell size. In this case, one or a few such files should
suffice, as the cells are expected to behave very similarly in all frames. 
There is an option here, which is mainly for example purposes, using the
provided dataset, and that is that a list of image files can be provided 
directly. These should then be in the form of a list of lists, each sublist 
being composed of matrices representing the colors. See example for useage.}

\item{frameNum}{This identifies which of the frames in the file that 
contains the information about the nuclear staining. In RGB files, this is 
often number 3.}

\item{numPix}{If the frames are very large, this can be used to reduce the 
computational burden. For reproducibility reasons, it might be clever to
run this multiple times, in this case, to bootstrap, or alternatively to
set a seed before starting.}

\item{numOfImgs}{If the provided files are nd2 format, they can contain
multiple files. In this case, this flag can be used to restrict the number of
used images.}

\item{returAllVals}{Should all the island diameters, including the very tiny,
noisy ones be returned, or should a focused value be returned?}
}
\value{
A size cutoff value in pixels, which corresponds to either:
 - the media of the "real" island diameters, i.e. after exclusion 
 of noisy, minimal islands of 1-10 pixel diameters.
 - a vector of all island diameters, including the noise, if returAllVals
 is set to TRUE.
}
\description{
This function is needed in situations where the average cell size is not 
known. What it does is that it identifies the average cell nucleus size
in a provided file. In this case, therefore, it is important that the frame
number identifies a frame with nuclear staining, such as DAPI or similar.
}
\examples{
See example in \code{\link{imageStatGen}}
}
