% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negImage.R
\docType{data}
\name{negImage}
\alias{negImage}
\title{A few CASPR2-negative cells}
\format{
A three-dimensional array representing red (IgG), green (GFP) and 
blue (DAPI) colors.
}
\usage{
data("negImage")
}
\description{
This is a negative control image, used for help functions and vignettes.
}
\keyword{datasets}
