% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posImage.R
\docType{data}
\name{posImage}
\alias{posImage}
\title{A GABAB-R red-blue non RGB TIFF image}
\format{
A list of two matrices, the first red (AF568), 
the second blue (DAPI).
}
\usage{
data("posImage")
}
\description{
This is a positive control image, which is one of three images
meant to be used for testing purposes in the package.
}
\keyword{datasets}
