% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posImage.R
\docType{data}
\name{posImage}
\alias{posImage}
\title{A few CASPR2-positive cells}
\format{
A three-dimensional array representing red (IgG), green (GFP) and
blue (DAPI) colors.
}
\usage{
data("posImage")
}
\description{
This is a positive control image, used for help functions and vignettes.
}
\keyword{datasets}
