% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveImage.R
\name{saveImage}
\alias{saveImage}
\title{Saving individually coloured png images from raw image input}
\usage{
saveImage(
  imgDirs,
  imgNames,
  frameNums = "All",
  frameCols,
  normTo = "bit",
  numBit = 16,
  outDir,
  numOfImgs = "All",
  numPix = "All"
)
}
\arguments{
\item{imgDirs}{The directory containing the image files. nd2, czi,
tiff and png formats or a list of three-dimensional arrays, where each 
layer in the third dimension represents a color are currently supported. 
nd2, czi or non-normalised, integer TIFF are clearly preferable.}

\item{imgNames}{The names of the images. This argument is used both for
naming of the rows in the output matrix, but also for naming of the images
if such are generated using the diagnoImgs flag.}

\item{frameNums}{This identifies which of the frames in the files that 
should be plotted. Defaults to "All", which then also generates a merge.}

\item{frameCols}{Which color does the frame/frames in question have? 
Strangely, in some file formats such as nd2, it varies - not all are 
"R-G-B", but "G-R-B" also occurs. Can take the values "R"/"Red", "G"/"Green",
and "B"/"Blue". Should be a vector the same length as frameNums. If nothing
is provided, it defaults to only saving the merge.}

\item{normTo}{This argument has three possible values: "bit", a 
numeric value, or "max". Vectors of individual values the same length as 
frameCols are also accepted. 
"bit" makes all pictures in a set comparable, but will make it hard
to see the subtle shifts often present in negative files. The value option is
often the most suitable, but requires some knowledge of the range of values 
expected in the file. For this, the \link{getImageIntensities} function can 
be useful to provide insights to provide reasonable values. If this value is
set to a lower value than the true max, the data will be truncated and a 
warning thrown. "max", the standard in png or tiff file formats, and will 
create non-comparable, often rather useless pictures, but it can be tried.}

\item{numBit}{How many bit are the nd2 files? Can be found using 
\code{\link[RBioFormats]{read.metadata}} function, in the 
"$coreMetadata$bitsPerPixel" slot. NB! If more than one image is present in
the nd2 file, one needs to refer to the first slot in the list, i.e., sub-
setting the metadata objects with double hard brackets.}

\item{outDir}{The directory that the images should be saved in. Subfolders
are created within this directory for each color.}

\item{numOfImgs}{If the provided files are nd2 format, they can contain
multiple files. In this case, this flag can be used to restrict the number of
used images.}

\item{numPix}{If the frames are very large, this can be used to reduce the 
computational burden. For reproducibility reasons, it might be clever to
run this multiple times, in this case, to bootstrap, or alternatively to
set a seed before starting. NB! If this command is used, then the images
will not be identical from round to round, and will not be comparable to the
output of other functions, as a random subset of the pictures are used.}
}
\value{
Named png images that are scaled to the max allowed value in the 
raw image file and coloured according to frameCol.
}
\description{
THis function simply generates pngs from raw image files, such as .nd2.
}
